requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
pandas
openpyxl
Flask-Testing